requirement.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]==4.7
pymongo
scikit-learn
pyaml
dill
dagshub
mlflow
dagshub
fastapi
uvicorn
-e .
